Add burst parameter to DropRateErrorModel (#134)
* Add burst parameter to DropRateErrorModel

This makes it so that the error model doesn't create loss bursts that
are longer than the burst parameter. This should prevent idle timeouts
when loss bursts are too long. It will however lower the overall loss
rate below what is configured, because we're not adjusting the drop rate
upward to compensate for breaking loss bursts (by forcing a forward).

* Minimize diff

* Simplify

* Suggestions from @marten-seemann